+++ java-io-base/src/test/java/nbbrd/io/sys/ProcessReaderTest.java
@@ -4,37 +4,40 @@ import org.junit.jupiter.api.Test;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.charset.Charset;
 import java.util.UUID;
 
+import static java.nio.charset.StandardCharsets.UTF_8;
+import static nbbrd.io.sys.ProcessReader.newReader;
+import static nbbrd.io.sys.ProcessReader.readToString;
 import static org.assertj.core.api.Assertions.*;
 
 public class ProcessReaderTest {
 
+    @SuppressWarnings({"resource", "DataFlowIssue", "deprecation"})
     @Test
     public void testFactories() {
-        assertThatNullPointerException()
-                .isThrownBy(() -> ProcessReader.newReader((String[]) null));
+        assertThatNullPointerException().isThrownBy(() -> newReader((String[]) null));
+        assertThatNullPointerException().isThrownBy(() -> newReader((Process) null));
+        assertThatNullPointerException().isThrownBy(() -> readToString((String[]) null));
+        assertThatNullPointerException().isThrownBy(() -> readToString((Process) null));
 
-        assertThatNullPointerException()
-                .isThrownBy(() -> ProcessReader.newReader((Process) null));
-
-        assertThatNullPointerException()
-                .isThrownBy(() -> ProcessReader.readToString((String[]) null));
-
-        assertThatNullPointerException()
-                .isThrownBy(() -> ProcessReader.readToString((Process) null));
+        assertThatNullPointerException().isThrownBy(() -> newReader(UTF_8, (String[]) null));
+        assertThatNullPointerException().isThrownBy(() -> newReader(UTF_8, (Process) null));
+        assertThatNullPointerException().isThrownBy(() -> readToString(UTF_8, (String[]) null));
+        assertThatNullPointerException().isThrownBy(() -> readToString(UTF_8, (Process) null));
     }
 
     @Test
     public void testContent() throws IOException {
         switch (OS.NAME) {
             case WINDOWS:
-                assertThat(new File(ProcessReader.readToString("where", "where"))).exists();
+                assertThat(new File(readToString(Charset.defaultCharset(), "where", "where"))).exists();
                 break;
             case LINUX:
             case MACOS:
             case SOLARIS:
-                assertThat(new File(ProcessReader.readToString("which", "which"))).exists();
+                assertThat(new File(readToString(Charset.defaultCharset(), "which", "which"))).exists();
                 break;
         }
     }
@@ -44,14 +47,14 @@ public class ProcessReaderTest {
         switch (OS.NAME) {
             case WINDOWS:
                 assertThatExceptionOfType(EndOfProcessException.class)
-                        .isThrownBy(() -> ProcessReader.readToString("where", UUID.randomUUID().toString()))
+                        .isThrownBy(() -> readToString(Charset.defaultCharset(), "where", UUID.randomUUID().toString()))
                         .withMessageStartingWith("Invalid exit value")
                         .withNoCause()
                         .matches(ex -> ex.getExitValue() != 0)
                         .matches(ex -> !ex.getErrorMessage().isEmpty());
 
                 assertThatExceptionOfType(EndOfProcessException.class)
-                        .isThrownBy(() -> ProcessReader.readToString("where", "/Q", UUID.randomUUID().toString()))
+                        .isThrownBy(() -> readToString(Charset.defaultCharset(), "where", "/Q", UUID.randomUUID().toString()))
                         .withMessageStartingWith("Invalid exit value")
                         .withNoCause()
                         .matches(ex -> ex.getExitValue() != 0)
@@ -61,7 +64,7 @@ public class ProcessReaderTest {
             case MACOS:
             case SOLARIS:
                 assertThatExceptionOfType(EndOfProcessException.class)
-                        .isThrownBy(() -> ProcessReader.readToString("which", UUID.randomUUID().toString()))
+                        .isThrownBy(() -> readToString(Charset.defaultCharset(), "which", UUID.randomUUID().toString()))
                         .withMessageStartingWith("Invalid exit value")
                         .withNoCause()
                         .matches(ex -> ex.getExitValue() != 0);